Refactor: Move player device and context API testing logic to separate class.
